requirements.txt
stripe==2.43.0
aniso8601==7.0.0
ansible==2.8.7
astroid==2.3.3
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
Django==2.2.11
django-appconf==1.0.3
django-ckeditor==6.0.0
django-cors-headers==3.2.1
django-countries==5.5
django-filter==2.2.0
django-fsm==2.6.1
django-fsm-admin==1.2.4
django-fsm-log==1.6.2
django-inline-svg==0.1.1
django-nested-admin==3.2.4
django-rest-swagger==2.2.0
django-webpack-loader==0.6.0
djangorestframework==3.10.3
djangorestframework-camel-case-2==0.0.1
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.9
future==0.18.2
graphviz==0.13.2
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==19.2
pandas==1.4.2
parso==0.5.1
Pillow==8.4.0
pluggy==0.13.1
promise==2.2.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.4.3
pyparsing==2.4.6
python-jsonrpc-server==0.3.1
python-language-server==0.31.1
python-monkey-business==1.0.0
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
Rx==1.6.1
simplejson==3.17.0
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.0
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.7
whitenoise==4.1.4
wrapt==1.11.2
